# tmp_pricelabs_swagger_pretty.json
{
  "swagger": "2.0",
  "info": {
...
    "version": "1.0.0",
    "title": "PriceLabs API Integration",
    "contact": {
      "email": "support@pricelabs.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.pricelabs.co",
  "basePath": "/v1/integration/api",
...
  "schemes": [
    "https"
  ],
  "security": [
    {
      "integration_token": []
    },
    {
      "integration_name": []
    }
  ],
  "paths": {
...
    "/integration": {
      "post": {
        "tags": [
          "PMS"
        ],
        "summary": "Update integration information",
        "description": "This endpoint can update integration specific details such as refreshing the API token if needed, or adjusting information that allows Pricelabs to send notices/updates.\n\n<b><i>sync_url</i></b>: Endpoint where Pricelabs can send a pricing data to be synced\n<b><i>calendar_trigger_url</i></b>: Endpoint where PriceLabs requests complete listing calendar information\n<b><i>hook_url</i></b>: Endpoint where PriceLabs can send error notifications.\n\nUpdating these three url endpoints results in a test body being sent with only the json body { verify => true } and no headers and requires a 200 response in return in order to be stored.\n\n When you call <b>/integration</b>, please note that if <b><i>regenerate_token</i></b> is true, you will receive a new token in response which should be used to make subsequent API calls.\n\n<b>Delta Push</b>: If you wish to receive only delta updates for prices and restrictions, then you can do so by setting <b><i>delta_only</i></b> flag as true. Once the flag is set to true, you will start to receive only those prices and restrictions which have changed since the last sync. If at any point, you wish to pull full data, you can do so by calling the /get_prices endpoint with the attribute <b><i>delta_only</i></b> as true.\n\n<b>IMPORTANT</b>: The <i>features</i> array should be used to configure only those settings that the PMS intends to actively read from PriceLabs for all connected listings. \n\n  <i>Please note that if a setting is marked as <b>true</b>, it becomes visible to users in PriceLabs, who may enable it with the expectation that the PMS will consume and apply it. Therefore, make sure to enable only those settings that are actually supported and intended to be consumed from PriceLabs.</i>\n  \n  <b>Once the integration is live, any updates to the supported features must first be discussed with PriceLabs. This allows us to coordinate the release of new features to all users and ensure they function as expected.</b>\n\n\n\n",
        "operationId": "integration",
        "parameters": [
          {
            "in": "body",
            "name": "integration",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "integration": {
                  "$ref": "#/definitions/IntegrationData"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request was processed and returns a message holding new API if requested and successful, or a simple Success and an error message which is populated on errors. <i>Please note that this response will not contain the signed headers as with the other API call responses.</i>",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid integration API key value",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "integration_token": {
      "type": "apiKey",
      "name": "X-INTEGRATION-TOKEN",
      "in": "header",
      "description": "This is an api token. This is generated and provided by PriceLabs."
    },
    "integration_name": {
      "type": "apiKey",
      "name": "X-INTEGRATION-NAME",
      "in": "header",
      "description": "This is a name field for the pms. The name must be approved and provided by PriceLabs."
    }
  },
  "definitions": {
...
    "IntegrationData": {
      "type": "object",
      "required": [
        "sync_url",
        "calendar_trigger_url",
        "hook_url",
        "regenerate_token"
      ],
      "properties": {
        "sync_url": {
          "type": "string"
        },
        "calendar_trigger_url": {
          "type": "string"
        },
        "regenerate_token": {
          "type": "boolean",
          "default": true,
          "description": "If not sent, it will not regenerate token. If sent blank, it will regenerate the token by default."
        },
        "hook_url": {
          "type": "string"
        },
        "features": {
          "$ref": "#/definitions/FeaturesControl"
        }
      }
    },
...
    "FeaturesControl": {
      "type": "object",
      "properties": {
        "min_stay": {
          "type": "boolean"
        },
        "check_in": {
          "type": "boolean"
        },
        "check_out": {
          "type": "boolean"
        },
        "monthly_weekly_discounts": {
          "type": "boolean"
        },
        "extra_person_fee": {
          "type": "boolean"
        },
        "los_pricing": {
          "type": "boolean"
        },
        "delta_only": {
          "type": "boolean",
          "description": "If this flag is set to true, PriceLabs will push only delta to the sync_url. In other words, only those prices and restrictions will be pushed which have changed since the last sync. If this flag is set to false, then all prices and restrictions will be pushed."
        }
      }
    },
...
  }
}